Clean Code and Add Admin Authentication (#70)
* Spell checking and code adjustments

* Added token authentication to admin POST endpoints

* restructure admin authentication

+++ server.js
@@ -62,29 +62,49 @@ app.get('/notes', async (_, res) => {
 });
 
 app.post('/call/create', async (req, res) => {
-    await calls.createCall(pool, req.body);
-    const response_data = await calls.getTotalCalls(pool);
-    io.emit('calls', response_data);
-    res.send(response_data);
+    if (WEBSITE_ACCESS_TOKEN !== req.query.token) {
+        res.sendStatus(403);
+    }
+    else {
+        await calls.createCall(pool, req.body);
+        const response_data = await calls.getTotalCalls(pool);
+        io.emit('calls', response_data);
+        res.send(response_data);
+    }
 });
 
 app.post('/note/create', async (req, res) => {
-    const response_data = await notes.createNote(pool, req.body.note);
-    io.emit('notes', await notes.getNotes(pool));
-    res.send(response_data);
+    if (WEBSITE_ACCESS_TOKEN !== req.query.token) {
+        res.sendStatus(403);
+    }
+    else {
+        const response_data = await notes.createNote(pool, req.body.note);
+        io.emit('notes', await notes.getNotes(pool));
+        res.send(response_data);
+    }
 });
 
 app.post('/note/delete', async (req, res) => {
-    const response_data = await notes.deleteNote(pool, req.body.note);
-    io.emit('notes', await notes.getNotes(pool));
-    res.send(response_data);
+    if (WEBSITE_ACCESS_TOKEN !== req.query.token) {
+        res.sendStatus(403);
+    }
+    else {
+        const response_data = await notes.deleteNote(pool, req.body.note);
+        io.emit('notes', await notes.getNotes(pool));
+        res.send(response_data);
+    }
 });
 
 app.post('/mishap/create', async (req, res) => {
-    await mishap.createMishap(pool, req.body.mishap);
-    const response_data = await mishap.getTotalMishaps(pool);
-    io.emit('mishaps', response_data);
-    res.send(response_data);
+    if (WEBSITE_ACCESS_TOKEN !== req.query.token) {
+        res.sendStatus(403);
+    }
+    else {
+        await mishap.createMishap(pool, req.body.mishap);
+        const response_data = await mishap.getTotalMishaps(pool);
+        io.emit('mishaps', response_data);
+        res.send(response_data);
+    }
 });
 
 app.get('/mishap', async (_, res) => {
@@ -101,7 +121,7 @@ app.post('/dispatch', (req, res) => {
     if (HERALD_TOKEN !== req.query.token) {
         res.sendStatus(403);
     } else {
-        console.log('recieved herald dispatch');
+        console.log('received herald dispatch');
         io.emit('dispatch', req.body);
         res.send({ success: true });
     }
